Created components of sections of home page
